Refactor Breadcrumb to section
