+++ src/components/ui/Breadcrumb/index.tsx
@@ -1,3 +1,2 @@
 export { default } from './Breadcrumb'
-export { default as BreadcrumbWrapper } from './BreadcrumbWrapper'
 export type { BreadcrumbProps } from './Breadcrumb'
